Create factory for Case and Document
