+++ Models/src/main/java/ev/projects/models/CaseFactory.java
@@ -1,0 +1,13 @@
+package ev.projects.models;
+
+public class CaseFactory {
+
+    public static Case createCase(String title, String description, User user) {
+        Case aCase = new Case();
+        aCase.setTitle(title);
+        aCase.setDescription(description);
+        aCase.setCreatorUser(user);
+        return aCase;
+    }
+
+}
